% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_diario.R
\name{tc_diario}
\alias{tc_diario}
\title{Recupera el tipo de cambio del día solicitado. Parametros: (fecha)
 Tipo de cambio diario}
\usage{
tc_diario(fecha)
}
\arguments{
\item{fecha}{date}
}
\value{
double tipo de cambio de la fecha solicitado.
}
\description{
Recupera el tipo de cambio del día solicitado. Parametros: (fecha)
}
\examples{
 tc_diario('2022-09-29')

}
