% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_mensual.R
\name{tc_mensual}
\alias{tc_mensual}
\title{Recupera el detalle completo de la base de tipos de cambio para el mes solicitado. Parametros: (Año, Mes)}
\usage{
tc_mensual(Anio, Mes)
}
\arguments{
\item{Anio}{numeric año}

\item{Mes}{numeric (1-12)}
}
\value{
dataframe detalle completo de lostipos de cambio para el mes solicitado.
}
\description{
Recupera el detalle completo de la base de tipos de cambio para el mes solicitado. Parametros: (Año, Mes)
}
\examples{
 tc_mensual(Anio=2019,Mes=9)
}
